requirements.txt
# Direct dependencies are discord.py, Pebble, and python-dotenv.

aiohttp==3.8.3
dateparser==1.1.4
discord.py==2.0.1
openai==0.25.0
Pebble==5.0.2
python-dotenv==0.21.0
